[FIX][18.0] plm - fix depend warning
Fixes the following warning:
UserWarning: Field 'product.template.product_variant_id' in dependency
of product.template.linkeddocuments should be searchable. This is
necessary to determine which records to recompute when
product.product.linkeddocuments is modified. You should either make the
field searchable, or simplify the field dependency.

product.template.product_variant_id is a non-stored computed field, so
it can't be used as an (implict) dependency in a related field.

product.template.product_variant_ids OTOH can be used as dependency.